menu changes, fixed bank interest, end turn bug
I've learned a lot about building neat menus due to another project I'm working on. There is now a custom menu in the game and it no longer relies on console-menu to handle player choices and menu rendering.

The function that handles resetting config.json for a new game had the wrong values for player and bank interest. They were set to 0.5 and 0.7 respectively. Which means for every turn the player gained 50 percent interest on their money(would be nice) and loans were dished out at 70 percent interest(not very nice).

Lastly, an errant copy-paste broke the End Turn function which pretty much caused the game to be unplayable. This has been fixed.